Further separates SSO and Admin modules
